Added some section banners to the library code
Since this library code base is supposed to be a single header/source,
it's very easy to get the code to a point that's impossible to browse
through without a symbol navigator.

To aid in browsing and also great for minimaps, section banners have
been added to key areas of the code base.